Conditionally install language bindings with `install-lib` target
Prior to this change, all library bindings would be installed with
`install-lib`. This is not desired with swig installs that want to
distribute a more stripped down version of the swig package.

Conditionally install the files based on the command line arguments
provided to configure, e.g., honor `--without-tcl` for Tcl, etc.

Closes: #1611
Signed-off-by: Enji Cooper <yaneurabeya@gmail.com>