Reformat tag lists in doxygen documentation
Reformat the tables of unsupported and ignored tags.  Previous table
format was not maintainable.  Changing to use multi-column lists, so
that entries can be easily sorted and updated.  The displayed order is
now by row instead of by column, which is decidedly easier to read.
Other than removing duplicates, no change to content.