Added a octave test for the concatenation operator.
Swig types can be used by the operators [a,b] or [a;b] if an octave
overload functions horzcat.m or vertcat.m are added in a @swig_ref
subdirectory.
This requires that is_object() (isobject() for octave >= 4.4) returns
true for swig types which is ensured by the swig octave runtime.