Update CSharp.html
global::System seems to be needed inside the typemap examples and also in the examples of generated code. But in these *example* C# programs, that would not be typical usage because the System namespace is almost always imported and then used implicitly. 

Also, it was already not used in the first of the Console... calls in the example. So this is more consistent.