Ruby: Add check for lower bounds of unsigned long (long) parameters
This check is missing in the implementation of the used Ruby C API
funtions and therefore is added as an explicit check by this commit.

The added check was intentionally placed inside the existing 'if' blocks
in order to minimize the potential of returning an OverflowError where a
TypeError was returned before this change.

Reference to Ruby issue: https://bugs.ruby-lang.org/issues/9089