Stop using Travis CI which doesn't work any longer
Unfortunately Travis CI can't be used for building open source projects
any longer in practice, as extra credits must be manually requested
every time from Travis support, which is just infeasible, as this would
need to be done every couple of days if not more often, and support
reply time is anything from a couple of days to more than a month.

So stop even trying to use it, as it doesn't work anyhow, and switch to
using GitHub Actions only.