Install rvm in CI setup script if it's not available
It is not available out of the box in GitHub environment, unlike under
Travis CI, so install it ourselves.

Note that we do it manually because installing the rael-gc/rvm PPA
doesn't seem to work with weird failures in GitHub Actions environment.

We also can't use gpg2 --recv-keys in this environment as it doesn't
work, apparently due to a firewall.