Don't bother building out of date PR versions on AppVeyor
Exit early if there is version of the PR newer than the one that we are
about to start building, to avoid wasting 3.5 hours of AppVeyor build
time on the outdated PR versions.

The snippet doing this was copied from

https://github.com/appveyor/ci/issues/38#issuecomment-70628826

(thanks to Tony Kelman). An alternative to doing this could be enabling
the "Rolling builds" option in the AppVeyor UI, see

https://www.appveyor.com/docs/build-configuration/#rolling-builds

but this method seems to work well in practice and is more explicit.