python/cast_mode(long long): Test for overflows.
This patch extends the unit tests to check for overflow exception if max+1 or min-1 is applied.

It also changes the behavior for doubles:
float(2**53) can be represented exactly, however it cannot be separated from float(2**53+1).
The old code accepted this edge case, the new code reject it to fullify the max+1 error test.